rename services to lib, remove unused code in createTeam
