+++ lib/teams-api.js
@@ -15,17 +15,11 @@ export async function getTeams () {
 }
 
 export async function createTeam (data) {
-  const res = await fetch(URL, {
+  return fetch(URL, {
     method: 'POST',
     body: JSON.stringify(data),
     headers: {
       'Content-Type': 'application/json; charset=utf-8'
     }
   })
-
-  if (res.status === 200) {
-    await this.refreshTeams()
-  } else {
-    throw new Error('Could not create new team')
-  }
 }
